fix: allow decimal values in quotes and compensation fields
updates the submission drawer and listing builder to accept decimal numbers up to 4 decimal places for quotes and compensation fields. changes the database schema from int to float for ask, minRewardAsk, and maxRewardAsk fields. removes the integer constraint from zod validation and rewrites the token input component to properly handle decimal input with formatting.

closes #1250

+++ src/features/listings/utils/submissionFormSchema.ts
@@ -31,7 +31,7 @@ const submissionSchema = (
         ])
         .optional(),
       otherInfo: z.string().optional(),
-      ask: z.union([z.number().int().min(0), z.null()]).optional(),
+      ask: z.union([z.number().min(0), z.null()]).optional(),
       eligibilityAnswers: z
         .array(
           z.object({
